fix clippy lints

+++ src/ffi.rs
@@ -1,4 +1,3 @@
-use libc;
 use std::ptr;
 
 #[allow(non_camel_case_types)]
@@ -13,7 +12,7 @@ mod internal {
 pub fn time() -> time_t {
     let retrieved_time = unsafe { internal::time(ptr::null_mut()) };
 
-    if retrieved_time == !(0 as u32) {
+    if retrieved_time == !0_u32 {
         panic!("Could not get time from system");
     }
